dashboard_etl_monitor/requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
google-cloud-bigquery>=3.11.0
